added ssh config

+++ helper/pacaur.sh
@@ -14,11 +14,13 @@ install_pacaur() {
   package_install expac
   package_install yajl
 
-  git_clone "https://aur.archlinux.org/cower.git" "$HOME/.aur/cower"
-  eval_cmd "Install cower" "cd $HOME/.aur/cower && makepkg PKGBUILD --skippgpcheck --install --noconfirm"
+  if ! cmd_exists pacaur; then
+    git_clone "https://aur.archlinux.org/cower.git" "$HOME/.aur/cower"
+    eval_cmd "Install cower" "cd $HOME/.aur/cower && makepkg PKGBUILD --skippgpcheck --install --noconfirm"
 
-  git_clone "https://aur.archlinux.org/pacaur.git" "$HOME/.aur/pacaur"
-  eval_cmd "Install pacaur" "cd $HOME/.aur/pacaur && makepkg PKGBUILD --install --noconfirm"
+    git_clone "https://aur.archlinux.org/pacaur.git" "$HOME/.aur/pacaur"
+    eval_cmd "Install pacaur" "cd $HOME/.aur/pacaur && makepkg PKGBUILD --install --noconfirm"
+  fi
 }
 
 pacaur_update() {
